chore: CI validates themes using the lazygit config schema
The latest version of the schema is used. This means the color schemes
in this repository are implicitly meant to be used in the latest lazygit
development version.

The schema version could also be locked to a recent release, but at this
time there is no nice way of resolving the latest version in the
validation logic (github action).